refactor: 재활용을 위한 ChipComponent 생성
